changed a type of the field optionProfitSum

+++ src/main/java/ru/tn/profitcalculator/service/calculator/ProductCalculateResult.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Builder
 public class ProductCalculateResult {
     private BigDecimal profitSum;
-    private BigDecimal optionProfitSum;
+    private Long optionProfitSum;
     private BigDecimal totalSum;
     private BigDecimal maxRate;
     private Integer daysCount;
